Split datastore renamed
